moved bms functions to module
